ml_models/tts_server/app.py
from flask import Flask, request, send_file, jsonify
import piper
import io
import os

app = Flask(__name__)

# Load Piper Voice
# Requires model file (.onnx) and config (.json)
# For now we'll assume they are mounted or downloaded
VOICE_MODEL = os.environ.get("VOICE_MODEL", "en_US-lessac-medium")
# In a real setup, you'd verify files exist

voice = None

print(f"Loading Piper voice: {VOICE_MODEL}...")
# Note: Python bindings for Piper can be tricky to install in generic containers
# We'll stub this if library loading fails, or assume the Dockerfile handles it correctly.
try:
    # This assumes the model file exists at ./models/<VOICE_MODEL>.onnx
    # We will wrap this in a try-except to avoid crash if model is missing
    # In a full implementation, you'd download the model on startup if missing.
    pass 
    # voice = piper.PiperVoice.load(f"./models/{VOICE_MODEL}.onnx") # Commented out until files exist
    print("Piper voice placeholder ready.")
except Exception as e:
    print(f"Error loading Piper voice: {e}")

@app.route('/synthesize', methods=['POST'])
def synthesize():
    text = request.json.get('text', '')
    if not text:
        return jsonify({'error': 'No text provided'}), 400
    
    # Placeholder for actual synthesis until model files are present
    # In reality:
    # audio_bytes = io.BytesIO()
    # voice.synthesize(text, audio_bytes)
    # audio_bytes.seek(0)
    # return send_file(audio_bytes, mimetype='audio/wav')

    return jsonify({'status': 'mock_success', 'message': 'TTS Engine ready but model files missing'}), 200

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5004)
