ml_models/tts_server/requirements.txt
flask==3.0.0
piper-tts==1.2.0
flask-cors==4.0.0
